requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
curlify==3.0.0
decouple==0.0.7
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-filter==25.1
django-jazzmin==2.6.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
facebook_business==23.0.1
flower==2.0.1
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
humanize==4.13.0
idna==3.10
kombu==5.5.4
multidict==6.6.4
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.20.1
